requirements.txt
openai
Pillow
easyocr
gradio
deta
